pyproject.toml
[project]
name = "dandelion-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Fede Barcelona", email = "fede_rico_94@hotmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.5",
    "uvicorn[standard]>=0.32.1",
    "jinja2>=3.1.4",
    "pymetasploit3==1.0.6"
]

[project.scripts]
dandelion-python = "dandelion_python:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
